backend/requirements.txt
asgiref==3.6.0
celery==5.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==4.0.1
dj-database-url
django-celery-results==2.5.1
django-cors-headers==4.0.0
django-heroku==0.3.1
django-model-utils==5.0.0
django-redis==5.3.0
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
langchain==0.3.18
langchain-chroma==0.1.4
langchain-community
langchain-core==0.3.34
langchain-openai==0.2.14
langchain-text-splitters==0.3.6
openai==1.58.1
httpx==0.28.1
httpx-sse==0.4.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
redis==4.6.0
rq==1.16.1
sqlparse==0.4.2
uvicorn==0.23.2
fastapi==0.115.6
boto3
botocore
gunicorn
cohere
chromadb[pg]
pgvector
scipy
